[Improve] Move QuicklyChecklist to quicklychecklist modules
