+++ features/quicklychecklist/impl/src/main/java/wottrich/github/io/smartchecklist/quicklychecklist/domain/ConvertQuicklyChecklistIntoJsonUseCase.kt
@@ -3,7 +3,7 @@ package wottrich.github.io.smartchecklist.quicklychecklist.domain
 import com.google.gson.Gson
 import wottrich.github.io.smartchecklist.coroutines.KotlinResultUseCase
 import wottrich.github.io.smartchecklist.coroutines.base.Result
-import wottrich.github.io.smartchecklist.datasource.data.model.QuicklyChecklist
+import wottrich.github.io.smartchecklist.quicklychecklist.domain.data.QuicklyChecklist
 
 class ConvertQuicklyChecklistIntoJsonUseCase : KotlinResultUseCase<QuicklyChecklist, String>() {
     override suspend fun execute(params: QuicklyChecklist): Result<String> {
